a little file cleanup and touching up the variable use in lampfadein
